+++ src/main/java/insane96mcp/survivalreimagined/event/SREventFactory.java
@@ -8,6 +8,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.FallingBlockEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.MinecraftForge;
 
@@ -47,6 +48,12 @@ public class SREventFactory {
         PlayerExhaustionEvent event = new PlayerExhaustionEvent(player, amount);
         MinecraftForge.EVENT_BUS.post(event);
         return event.getAmount();
+    }
+
+    public static void onCakeEatEvent(Player player, BlockPos pos, LevelAccessor level)
+    {
+        CakeEatEvent event = new CakeEatEvent(player, pos, level);
+        MinecraftForge.EVENT_BUS.post(event);
     }
 
     public static void onFallingBlockLand(FallingBlockEntity fallingBlock)