+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 3.10.2
+* Each cake bite now restores 40% of missing health
 * Disabled Golden Absorption. Gold armor and shield now grant +2 max health
 * Invincibility frames are now synced when hit by arrows
 * Glass can now break if entities fall on it